Delete unused code.
